Windows/requirements.txt
keyboard
g4f
python-docx